requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Supabase - using latest compatible versions
supabase>=2.4.0
websockets>=12.0

# Environment and Config
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# HTTP requests
requests==2.31.0
httpx[http2]>=0.26,<0.29

# Background tasks
rq[redis]==1.15.1
redis==5.0.1

# ML/CV
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0

# OCR and Content Moderation (Day 6)
pytesseract==0.3.10
easyocr==1.7.0
pillow>=9.0.0
regex==2023.10.3
profanity-check==1.0.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

shapely==2.0.4

pytesseract==0.3.10